renamed util to utils. renamed dimension module to nmr_method
