+++ .flake8
@@ -2,9 +2,6 @@
 ignore = C901
 exclude = .eggs,*.egg,build
 filename = *.pyx,*.px*, *py
-max-line-length = 90
+max-line-length = 88
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
-
-
-## E203, E266, E501, F403, F811, F841, F401, C901, E741, W503, C901 C901 E722
