+++ app/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+__version__ = "v0.2.0a0"
