+++ app/assets/media.css
@@ -112,7 +112,7 @@
   .my-card {
     border-radius: 0;
   }
-  #interface-dimension-spin-system {
+  #interface-nmr_method-spin-system {
     position: fixed;
     object-position: right;
     top: 0;
