+++ app/nmr_method/post_simulation_widgets.py
