+++ app/nmr_method/simulation_widgets.py
