+++ app/nmr_method/utils.py
@@ -35,7 +35,8 @@ def update_method_info(methods):
                             #     "X",
                             #     color="danger",
                             #     block=True,
-                            #     id={"type": "initiate-remove-spin-system", "index": i},
+                            #     id={"type": "initiate-remove-spin-system",
+                            #       "index": i},
                             # ),
                         ]
                     ),
